Move additional components into compose module and Utils into utils module
