+++ app/src/main/java/de/schnettler/scrobbler/screens/details/ArtistDetails.kt
@@ -14,14 +14,14 @@ import androidx.compose.ui.unit.dp
 import de.schnettler.database.models.EntityWithStats.TrackWithStats
 import de.schnettler.database.models.EntityWithStatsAndInfo.ArtistWithStatsAndInfo
 import de.schnettler.scrobbler.R
-import de.schnettler.scrobbler.components.Carousel
+import de.schnettler.scrobbler.ui.common.compose.Carousel
 import de.schnettler.scrobbler.components.ChipRow
 import de.schnettler.scrobbler.components.CollapsingToolbar
 import de.schnettler.scrobbler.components.ExpandingInfoCard
 import de.schnettler.scrobbler.components.ListWithTitle
-import de.schnettler.scrobbler.components.ListeningStats
-import de.schnettler.scrobbler.components.MediaCard
-import de.schnettler.scrobbler.components.rememberDominantColorCache
+import de.schnettler.scrobbler.ui.common.compose.ListeningStats
+import de.schnettler.scrobbler.ui.common.compose.MediaCard
+import de.schnettler.scrobbler.ui.common.compose.rememberDominantColorCache
 import de.schnettler.scrobbler.ui.common.compose.PlainListIconBackground
 import de.schnettler.scrobbler.ui.common.compose.UIAction
 import de.schnettler.scrobbler.ui.common.compose.UIAction.ListingSelected
