+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/DynamicTheming.kt
@@ -1,4 +1,4 @@
-package de.schnettler.scrobbler.components
+package de.schnettler.scrobbler.ui.common.compose
 
 import android.content.Context
 import android.util.LruCache
