+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/Header.kt
@@ -1,4 +1,4 @@
-package de.schnettler.scrobbler.components
+package de.schnettler.scrobbler.ui.common.compose
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.ExperimentalAnimationApi
@@ -18,7 +18,6 @@ import androidx.compose.ui.layout.WithConstraints
 import androidx.compose.ui.platform.AmbientDensity
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.min
-import de.schnettler.scrobbler.util.Orientation
 
 @OptIn(ExperimentalAnimationApi::class)
 @Composable
