+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/MediaCard.kt
@@ -1,4 +1,4 @@
-package de.schnettler.scrobbler.components
+package de.schnettler.scrobbler.ui.common.compose
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
@@ -29,7 +29,6 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import de.schnettler.scrobbler.ui.common.compose.theme.AppColor
 import de.schnettler.scrobbler.ui.common.util.abbreviate
-import de.schnettler.scrobbler.util.Orientation
 import dev.chrisbanes.accompanist.coil.CoilImage
 
 @Composable
