+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/Spacer.kt
@@ -1,11 +1,10 @@
-package de.schnettler.scrobbler.components
+package de.schnettler.scrobbler.ui.common.compose
 
 import androidx.compose.foundation.layout.preferredHeight
 import androidx.compose.foundation.layout.preferredWidth
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.Dp
-import de.schnettler.scrobbler.util.Orientation
 
 @Composable
 fun Spacer(size: Dp, orientation: Orientation = Orientation.Vertical) {
@@ -13,4 +12,9 @@ fun Spacer(size: Dp, orientation: Orientation = Orientation.Vertical) {
         Orientation.Vertical -> androidx.compose.foundation.layout.Spacer(modifier = Modifier.preferredHeight(size))
         Orientation.Horizontal -> androidx.compose.foundation.layout.Spacer(modifier = Modifier.preferredWidth(size))
     }
+}
+
+enum class Orientation {
+    Vertical,
+    Horizontal
 }
