+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/StatsRow.kt
@@ -1,4 +1,4 @@
-package de.schnettler.scrobbler.components
+package de.schnettler.scrobbler.ui.common.compose
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
