+++ ui/common/util/src/main/java/de/schnettler/scrobbler/ui/common/util/Resources.kt
@@ -1,4 +1,4 @@
-package de.schnettler.scrobbler.util
+package de.schnettler.scrobbler.ui.common.util
 
 const val AUTH_ENDPOINT = "https://www.last.fm/api/auth/"
 const val REDIRECT_URL = "de.schnettler.scrobble://auth"
